fix: simple key handling upto 1024 chars.
Currently, if a key is >128 characters, it is
handled as a complex key mapping. This renders some
OpenAPI spec paths invalid.
Until the same PR https://github.com/go-yaml/yaml/pull/1037
is not accepted in the upstream repository,
we are adding the changes in this fork and will be
using this in deck.

Fixes: https://github.com/Kong/go-apiops/issues/198